requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
pandas==2.2.0
openpyxl==3.1.2

# Embeddings & Vector Store
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.3

# LLM Integration
requests==2.31.0
openai==1.10.0

# Frontend
streamlit==1.30.0

# Utilities
python-dotenv==1.0.0
